--- pkg/executor/join/BUILD.bazel
@@ -1,108 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "join",
-    srcs = [
-        "base_join_probe.go",
-        "concurrent_map.go",
-        "hash_join_base.go",
-        "hash_join_v1.go",
-        "hash_join_v2.go",
-        "hash_table_v1.go",
-        "hash_table_v2.go",
-        "index_lookup_hash_join.go",
-        "index_lookup_join.go",
-        "index_lookup_merge_join.go",
-        "inner_join_probe.go",
-        "join_row_table.go",
-        "joiner.go",
-        "merge_join.go",
-        "outer_join_probe.go",
-    ],
-    importpath = "github.com/pingcap/tidb/pkg/executor/join",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/executor/aggregate",
-        "//pkg/executor/internal/applycache",
-        "//pkg/executor/internal/exec",
-        "//pkg/executor/internal/vecgroupchecker",
-        "//pkg/executor/unionexec",
-        "//pkg/expression",
-        "//pkg/parser/mysql",
-        "//pkg/parser/terror",
-        "//pkg/planner/core",
-        "//pkg/sessionctx",
-        "//pkg/sessionctx/stmtctx",
-        "//pkg/sessionctx/variable",
-        "//pkg/types",
-        "//pkg/util",
-        "//pkg/util/bitmap",
-        "//pkg/util/channel",
-        "//pkg/util/chunk",
-        "//pkg/util/codec",
-        "//pkg/util/collate",
-        "//pkg/util/dbterror/exeerrors",
-        "//pkg/util/disk",
-        "//pkg/util/execdetails",
-        "//pkg/util/hack",
-        "//pkg/util/logutil",
-        "//pkg/util/memory",
-        "//pkg/util/mvmap",
-        "//pkg/util/ranger",
-        "//pkg/util/serialization",
-        "//pkg/util/sqlkiller",
-        "//pkg/util/syncutil",
-        "@com_github_pingcap_errors//:errors",
-        "@com_github_pingcap_failpoint//:failpoint",
-        "@org_uber_go_zap//:zap",
-    ],
-)
-
-go_test(
-    name = "join_test",
-    timeout = "short",
-    srcs = [
-        "bench_test.go",
-        "concurrent_map_test.go",
-        "hash_table_v1_test.go",
-        "hash_table_v2_test.go",
-        "index_lookup_join_test.go",
-        "index_lookup_merge_join_test.go",
-        "inner_join_probe_test.go",
-        "join_row_table_test.go",
-        "join_stats_test.go",
-        "joiner_test.go",
-        "left_outer_join_probe_test.go",
-        "merge_join_test.go",
-        "right_outer_join_probe_test.go",
-        "row_table_builder_test.go",
-    ],
-    embed = [":join"],
-    flaky = True,
-    shard_count = 48,
-    deps = [
-        "//pkg/config",
-        "//pkg/domain",
-        "//pkg/executor/internal/testutil",
-        "//pkg/expression",
-        "//pkg/parser/ast",
-        "//pkg/parser/mysql",
-        "//pkg/planner/core",
-        "//pkg/session",
-        "//pkg/sessionctx",
-        "//pkg/sessionctx/variable",
-        "//pkg/testkit",
-        "//pkg/types",
-        "//pkg/util",
-        "//pkg/util/chunk",
-        "//pkg/util/codec",
-        "//pkg/util/disk",
-        "//pkg/util/hack",
-        "//pkg/util/memory",
-        "//pkg/util/mock",
-        "//pkg/util/sqlkiller",
-        "@com_github_pingcap_errors//:errors",
-        "@com_github_pingcap_failpoint//:failpoint",
-        "@com_github_stretchr_testify//require",
-    ],
-)
